Move PropertiesAsset.java to fattest.simplicity
